refactor(EST-2132-FE): rename more services/models to kebab-case
