+++ frontend/src/app/features/settings-dialog/components/embedding-models-tab/create-embedding-model-modal/create-embedding-model-modal.component.ts
@@ -8,7 +8,7 @@ import { AppIconComponent } from '../../../../../shared/components/app-icon/app-
 import { CustomInputComponent } from '../../../../../shared/components/form-input/form-input.component';
 import { ButtonComponent } from '../../../../../shared/components/buttons/button/button.component';
 import { ToggleSwitchComponent } from '../../../../../shared/components/form-controls/toggle-switch/toggle-switch.component';
-import { LLM_Provider } from '../../../models/LLM_provider.model';
+import { LLM_Provider } from '../../../models/llm-provider.model';
 import { EmbeddingModel } from '../../../models/embeddings/embedding.model';
 import { EmbeddingModelsService } from '../../../services/embeddings/embeddings.service';
 import { getProviderIconPath } from '../../../utils/get-provider-icon';
