+++ frontend/src/app/features/settings-dialog/components/embedding-models-tab/model-selector-modal/model-selector-modal.component.ts
@@ -14,9 +14,9 @@ import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { finalize, forkJoin } from 'rxjs';
 
 import { AppIconComponent } from '../../../../../shared/components/app-icon/app-icon.component';
-import { LLM_Provider, ModelTypes } from '../../../models/LLM_provider.model';
+import { LLM_Provider, ModelTypes } from '../../../models/llm-provider.model';
 import { EmbeddingModel } from '../../../models/embeddings/embedding.model';
-import { LLM_Providers_Service } from '../../../services/LLM_providers.service';
+import { LLM_Providers_Service } from '../../../services/llm-providers.service';
 import { EmbeddingModelsService } from '../../../services/embeddings/embeddings.service';
 import { getProviderIconPath } from '../../../utils/get-provider-icon';
 import { AllModelsModalComponent } from '../all-models-modal/all-models-modal.component';
