+++ frontend/src/app/features/settings-dialog/components/llm-models-tab/model-selector-modal/model-selector-modal.component.ts
@@ -15,9 +15,9 @@ import { forkJoin } from 'rxjs';
 import { finalize } from 'rxjs/operators';
 
 import { AppIconComponent } from '../../../../../shared/components/app-icon/app-icon.component';
-import { LLM_Provider, ModelTypes } from '../../../models/LLM_provider.model';
+import { LLM_Provider, ModelTypes } from '../../../models/llm-provider.model';
 import { LLM_Model } from '../../../models/llms/LLM.model';
-import { LLM_Providers_Service } from '../../../services/LLM_providers.service';
+import { LLM_Providers_Service } from '../../../services/llm-providers.service';
 import { LLM_Models_Service } from '../../../services/llms/LLM_models.service';
 import { getProviderIconPath } from '../../../utils/get-provider-icon';
 import { AllModelsModalComponent, AllModelsResult } from '../all-models-modal/all-models-modal.component';
