+++ frontend/src/app/features/settings-dialog/models/llm-provider.model.ts
