+++ frontend/src/app/features/settings-dialog/services/realtime-llms/full-reamtime-config.service.ts
@@ -4,8 +4,8 @@ import { map } from 'rxjs/operators';
 
 import { RealtimeModelsService } from './real-time-models.service';
 import { RealtimeModelConfigsService } from './real-time-model-config.service';
-import { LLM_Providers_Service } from '../LLM_providers.service';
-import { LLM_Provider, ModelTypes } from '../../models/LLM_provider.model';
+import { LLM_Providers_Service } from '../llm-providers.service';
+import { LLM_Provider, ModelTypes } from '../../models/llm-provider.model';
 import { RealtimeModelConfig } from '../../models/realtime-voice/realtime-llm-config.model';
 import { RealtimeModel } from '../../models/realtime-voice/realtime-model.model';
