+++ frontend/src/app/features/tools/models/tool-config.model.ts
