+++ frontend/src/app/services/full-tool-config.service.ts
@@ -7,7 +7,7 @@ import { ToolConfigService } from './tool-config.service';
 import {
   CreateToolConfigRequest,
   GetToolConfigRequest,
-} from '../features/tools/models/tool_config.model';
+} from '../features/tools/models/tool-config.model';
 
 export interface FullToolConfig extends Tool {
   toolConfigs: GetToolConfigRequest[]; // List of related tool configurations
