+++ frontend/src/app/user-settings-page/tools/tool-configuration-dialog/tool-config-list/tool-config-list.component.ts
@@ -7,7 +7,7 @@ import {
   SimpleChanges,
   ChangeDetectionStrategy,
 } from '@angular/core';
-import { ToolConfig } from '../../../../features/tools/models/tool_config.model';
+import { ToolConfig } from '../../../../features/tools/models/tool-config.model';
 import { NgClass, NgFor, NgIf } from '@angular/common';
 
 @Component({
