+++ frontend/src/app/visual-programming/components/flow-graph-context-menu/tools-menu/tools-menu.component.ts
@@ -9,7 +9,7 @@ import {
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ToolConfigService } from '../../../../services/tool-config.service';
-import { ToolConfig } from '../../../../features/tools/models/tool_config.model';
+import { ToolConfig } from '../../../../features/tools/models/tool-config.model';
 import { NodeType } from '../../../core/enums/node-type';
 
 @Component({
